Explain in Readme Error identification
The purpose of this fork is to Identify what/where and WHY their is an error for this package. The issue is from the new authentication policy from Netflix. The API is no longer accessible, if not authenticated.
Possible clues and tracks to explored are detailed in the description.
This repo (for now) Doesn't offer a ready to use solution.